Add: ✨ Why it's a good idea to avoid heights
In Default, CSS wants to be responsive for the website. If you
did something on the CSS, and  made it non-responsive, then it is you
that made the issue.

If you truly wants to have height on an element. Just don't in default
try to play with the height of the element, instead use padding instead